fix: add legacy selector for edit page
